Analog read and send value to app example added
Also removed plab.pde files as they interferred with processing
development interface

Arduino:
The example reads an analog value from the A0 input. It converts this
value to a string. It sends the string to the app.

Processing:
The example gets string from bluetooth. It assumes string to be an integer
and translates it to integer. The integer value is written to screen, and
a needle shows how large it is on a scale from 0 to 1024.